+++ backend/products/views.py
@@ -26,11 +26,14 @@ def get_products(request):
 @permission_classes([IsAuthenticated])
 def update_product(request, pk):
     product = get_object_or_404(Product, pk=pk)
-    if request.method == 'PUT':
-        serializer = ProductSerializer(product, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data)
-    elif request.method == 'DELETE':
-        product.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    if product.user.id == request.user.id:
+        if request.method == 'PUT':
+            serializer = ProductSerializer(product, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data)
+        elif request.method == 'DELETE':
+            product.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
